\name{diff_exp_example1}
\alias{diff_exp_example1}
\docType{data}
\title{example of microarray data (data processed from GEO GSE9008)}
\description{
  example of microarray data (data processed from GEO GSE9008)
}
\usage{data(diff_exp_example1)}
\format{
  Data frames with 20861 observations on the following 3 variables.
  \describe{
    \item{\code{gene}}{a character vector}
    \item{\code{pvalue}}{a numeric vector}
    \item{\code{logFC}}{a numeric vector}
  }
}
\source{
Whyte L, Huang YY, Torres K, Mehta RG. Molecular mechanisms of resveratrol action in lung cancer cells using dual protein and microarray analyses. Cancer Res 2007.
}
\keyword{datasets}