\name{downloadAbsentFile}
\alias{downloadAbsentFile}
\title{
downloadAbsentFile
}
\description{
download a file only if it is not present.
}
\usage{
downloadAbsentFile(urlStr, oD = tempdir())
}
\arguments{
  \item{urlStr}{
url from which to download the file
}
\item{oD}{
directory where to store the file
}
}
\author{
Andrea Franceschini
}
