\name{downloadAbsentFileSTRING}
\alias{downloadAbsentFileSTRING}
\title{
downloadAbsentFileSTRING
}
\description{
download a STRING file only if it is not present or if it is corrupted.
}
\usage{
downloadAbsentFileSTRING(urlStr, oD = tempdir())
}
\arguments{
  \item{urlStr}{
url from which to download the file
}
\item{oD}{
directory where to store the file
}
}
\author{
Andrea Franceschini
}
