\name{get_annotations}
\alias{\S4method{get_annotations}{STRINGdb}}
\title{
get_annotations
}
\description{
Loads and returns STRING annotations (i.e. GO annotations, KEGG pathways, domain databases). 
The annotations are stored in the "annotations" variable.
}
\usage{
\S4method{get_annotations}{STRINGdb}( )
}
\value{
a data frame containing the annotations to the STRING proteins (e.g. GeneOntology, KEGG pathways, InterPro domains)
}
\author{
Andrea Franceschini
}


