\name{get_annotations_desc}
\alias{\S4method{get_annotations_desc}{STRINGdb}}
\title{
get_annotations_desc
}
\description{
Returns a data frame with the description of every STRING annotation term 
  (it downloads and caches the information the first time that is called).
}
\usage{
\S4method{get_annotations_desc}{STRINGdb}()
}
\value{
data frame with the description of every STRING annotation term.
}
\author{
Andrea Franceschini
}


