\name{get_proteins}
\alias{\S4method{get_proteins}{STRINGdb}}
\title{
get_proteins
}
\description{
Returns the STRING proteins data frame.
(it downloads and caches the information the first time that is called).
}
\usage{
\S4method{get_proteins}{STRINGdb}()
}
\value{
STRING proteins data frame.
}
\author{
Andrea Franceschini
}


