\name{get_pubmed}
\alias{\S4method{get_pubmed}{STRINGdb}}
\title{
get_pubmed
}
\description{
Returns vector with the PUBMED IDs of the publications that contain the names of the proteins in the input vector.
}
\usage{
\S4method{get_pubmed}{STRINGdb}(string_ids)
}
\arguments{
  \item{string_ids}{
a vector of STRING identifiers
}
}
\value{
vector with the PUBMED IDs of the publications that contain the names of the proteins in the input vector.
}
\author{
Andrea Franceschini
}


