\name{get_pubmed_interaction}
\alias{\S4method{get_pubmed_interaction}{STRINGdb}}
\title{
get_pubmed_interaction
}
\description{
Returns vector with the PUBMED IDs of the publications that contain the names of both the input proteins.
}
\usage{
\S4method{get_pubmed_interaction}{STRINGdb}(STRING_id_a, STRING_id_b )
}
\arguments{
  \item{STRING_id_a}{
STRING identifier
}
  \item{STRING_id_b}{
STRING identifier
}
}
\value{
vector with the PUBMED IDs of the publications that contain the names of both the input proteins.
}
\author{
Andrea Franceschini
}


