\name{get_STRING_species}
\alias{get_STRING_species}
\title{
get_STRING_species
}
\description{
Returns a data frame with the species (i.e. organisms) that are present in STRING.
}
\usage{
get_STRING_species(version=NULL, species_name=NULL)
}
\arguments{
  \item{version}{
STRING version
}
\item{species_name}{
name of the species that you are searching
}
}
\value{
data frame with the species (i.e. organisms) that are present in STRING.
}
\author{
Andrea Franceschini
}


