\name{get_subnetwork}
\alias{\S4method{get_subnetwork}{STRINGdb}}
\title{
get_subnetwork
}
\description{
Returns the subgraph generated by the given input proteins.
For a regulatory network, regulator-to-target edge directions are preserved.
}
\usage{
\S4method{get_subnetwork}{STRINGdb}(string_ids )
}
\arguments{
  \item{string_ids}{
a vector of STRING identifiers
}
}
\value{
Returns the subgraph (i.e. an iGraph object) generated by the given input proteins.
}
\author{
Andrea Franceschini
}

