\name{interactions_example}
\alias{interactions_example}
\docType{data}
\title{example of a protein-protein interactions sorted data frame}
\description{
  example of a sorted list of protein-protein interactions, resulta our cooccurrence algorithm (SVD_Phy)
}
\usage{data(interactions_example)}
\format{
  Data frames with 20861 observations on the following 3 variables.
  \describe{
    \item{\code{proteinA}}{a character vector}
    \item{\code{proteinB}}{a character vector}
    \item{\code{score}}{a numeric vector}
  }
}
\keyword{datasets}