\name{ppie.compLambdaL1L2}
\alias{ppie.compLambdaL1L2}

\title{
ppie.compLambdaL1L2 
}
\description{
Compute lambda L1L2
}
\usage{
ppie.compLambdaL1L2(degreesI, degreesJ, edgeNum)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{degreesI}{
vector containing the degrees of the nodes present in nodeSet I
}
  \item{degreesJ}{
vector containing the degrees of the nodes present in nodeSet J
}
  \item{edgeNum}{
total number of edges of the entire graph
}
}
\details{
compute the number of expected interactions between two sets of nodes
}
\author{
Andrea Franceschini
}
