\name{ppie.getNumEdgesBetween}
\alias{ppie.getNumEdgesBetween}

\title{
ppie.getNumEdgesBetween
}
\description{
find the number of interactions between two sets of nodes
}
\usage{
ppie.getNumEdgesBetween(graph, nodesFrom, nodesTo)
}

\arguments{
  \item{graph}{
igraph object{}
}
  \item{nodesFrom}{
list of nodes
}
  \item{nodesTo}{
list of nodes
}
}

\value{
return the number of interactions between two sets of nodes
}
\author{
Andrea Franceschini
}