\name{renameColDf}
\alias{renameColDf}
\title{
renameColDf
}
\description{
Rename a column of a data frame
}
\usage{
renameColDf(df, colOldName, colNewName)
}
\arguments{
  \item{df}{
input data frame
}
\item{colOldName}{
column name to be changed
}
\item{colNewName}{
new column name
}
}
\value{
data frame with the column name changed
}
\author{
Andrea Franceschini
}
