\name{get_bioc_graph}
\alias{\S4method{get_bioc_graph}{STRINGdb}}
\title{
get_bioc_graph
}
\description{
Returns the interaction graph as an object of the graph package in Bioconductor.
}
\usage{
\S4method{get_bioc_graph}{STRINGdb}()
}
\value{
interaction graph as an object of the graph package in Bioconductor.
}
\author{
Andrea Franceschini
}


