\name{get_neighbors}
\alias{\S4method{get_neighbors}{STRINGdb}}
\title{
get_neighbors
}
\description{
Get the neighborhoods of a protein (or of a vector of proteins) that is given in input.
}
\usage{
\S4method{get_neighbors}{STRINGdb}(string_ids)
}
\arguments{
  \item{string_ids}{
a vector of STRING identifiers
}
}
\value{
vector containing the neighborhoods of a protein (or of a vector of proteins) that is given in input.
}
\author{
Andrea Franceschini
}


